# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Learn Japanese API Documentation",
        "description": "API endpoints for Learn Japanese application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8081/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/quizzes/dynamic": {
            "get": {
                "tags": [
                    "Quizzes"
                ],
                "summary": "Generate Dynamic Quiz",
                "description": "Generate a dynamic multiple-choice quiz based on vocabulary.",
                "operationId": "b3de7fa2180791d2d20b5d2b45737036",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/auth/vocabularies/{id}/bookmark": {
            "post": {
                "tags": [
                    "User Progress"
                ],
                "summary": "Toggle Bookmark",
                "operationId": "0058e713a4dac7366e100fcfc97e19e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/vocabularies/{id}/status": {
            "post": {
                "tags": [
                    "User Progress"
                ],
                "summary": "Update Learn Status",
                "operationId": "c71df78c5c5cc18975d2c51efef7e1c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "mastered",
                                        "enum": [
                                            "learning",
                                            "mastered",
                                            "not_learned"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}